tcradix-atcount: workaround a compiler misoptimization.
If the function is looped (be it with while or goto), at -O1 or higher (but
not -O0 and -Og) its speed goes down by a factor of 5.  To avoid this, we can
copy the function so its first iteration is separate.  Somehow that's enough
despite the source code being identical to what we tail-jump.